backend/requirement.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
pymongo
websockets
